+++ app/services/scraper_orchestrator_service.rb
@@ -6,6 +6,7 @@ require "digest"
 # Orchestrates the complete web scraping workflow by coordinating URL validation,
 # HTTP fetching, HTML parsing, and data extraction services
 class ScraperOrchestratorService
+  include ScraperHelpers
   def self.call(url:, fields: [])
     new(
       url_validator: UrlValidatorService,
@@ -51,15 +52,6 @@ class ScraperOrchestratorService
     log_completion(start_time, false, false, error: error)
     wrapped_error = error.is_a?(ScraperErrors::BaseError) ? error : wrap_unexpected_error(error)
     error_response(wrapped_error)
-  end
-
-  def log_start_context(url, fields)
-    context = {
-      url: url,
-      fields_count: fields.size,
-      request_id: Thread.current[:request_id]
-    }
-    Rails.logger.info("ScraperOrchestrator started: #{context.to_json}")
   end
 
   private
@@ -114,49 +106,9 @@ class ScraperOrchestratorService
     result[:data]
   end
 
-  def convert_hash_to_fields_array(fields_hash)
-    fields_hash.map do |name, config|
-      field_config = config.is_a?(Hash) ? config.dup : { selector: config }
-      field_config[:name] = name.to_s
-      field_config
-    end
-  end
-
-  def extract_meta_tag(document, name)
-    meta_tag = document.at_css("meta[name='#{name}'], meta[property='#{name}']")
-    meta_tag&.attribute("content")&.value
-  end
-
   def cache_results(_url, _fields, _data)
     # Placeholder - will implement caching in a later step
     true
-  end
-
-  def cache_key(url, fields)
-    field_hash = Digest::SHA256.hexdigest(fields.sort.to_json)
-    "scraper:#{Digest::SHA256.hexdigest(url)}:#{field_hash}"
-  end
-
-  def success_response(data, cached:)
-    {
-      success: true,
-      data: data,
-      cached: cached,
-      error: nil
-    }
-  end
-
-  def error_response(error)
-    {
-      success: false,
-      data: nil,
-      cached: false,
-      error: error.message
-    }
-  end
-
-  def wrap_unexpected_error(error)
-    ScraperErrors::BaseError.new("Unexpected error: #{error.message}")
   end
 
   def log_completion(start_time, success, cached, error: nil)
